Django_Backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
web3==6.11.0
python-dotenv==1.0.0
Pillow==10.0.0
qrcode==7.4.2
pyzbar==0.1.9
